Add comments to indicate concept is For exposition only
Adding a concept to the standard is a serious endeavor because it can never be changed.
This softens the recommendation until we have consensus.

We believe there is merit to adding concept(s) for an adjancy_list and edgelist have merit because they
introduce new type not currently in the standard, namely a range of ranges compared to 1-dimensional
range for traditional containers.

The alternative is to unwind/expand the concept defintions and add the result in the requires clause.